+++ Monitor/rF2SMMonitor/VC16/rF2SMMonitor.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -46,6 +46,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\rF2SharedMemory\RF2SharedMemoryBuffers.cs" />
     <Compile Include="..\rF2SMMonitor\IniFile.cs" />
     <Compile Include="..\rF2SMMonitor\MainForm.cs">
       <SubType>Form</SubType>
@@ -56,9 +57,10 @@
     <Compile Include="..\rF2SMMonitor\Math.cs" />
     <Compile Include="..\rF2SMMonitor\Program.cs" />
     <Compile Include="..\rF2SMMonitor\Properties\AssemblyInfo.cs" />
-    <Compile Include="..\rF2SMMonitor\rF2Data.cs" />
+    <Compile Include="..\..\..\rF2SharedMemory\rF2Data.cs">
+      <Link>rF2SharedMemory\rF2Data.cs</Link>
+    </Compile>
     <Compile Include="..\rF2SMMonitor\TransitionTracker.cs" />
-    <Compile Include="..\rF2SMMonitor\RF2SharedMemoryReader.cs" />
     <EmbeddedResource Include="..\rF2SMMonitor\MainForm.resx">
       <DependentUpon>..\rF2SMMonitor\MainForm.cs</DependentUpon>
     </EmbeddedResource>
@@ -84,6 +86,7 @@
   <ItemGroup>
     <None Include="..\rF2SMMonitor\App.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
