+++ rF2SharedMemory/RF2SharedMemoryBuffers.cs
@@ -3,9 +3,9 @@ using System.IO;
 using System.IO.MemoryMappedFiles;
 using System.Runtime.InteropServices;
 using System.Threading;
-using rF2SMMonitor.rFactor2Data;
+using rF2SharedMemory.rFactor2Data;
 
-namespace rF2SMMonitor
+namespace rF2SharedMemory
 {
   public class MappedBuffer<MappedBufferT>
   {
@@ -145,7 +145,7 @@ namespace rF2SMMonitor
         var versionHeaderWithSize = new rF2MappedBufferVersionBlockWithSize();
         var versionHeader = new rF2MappedBufferVersionBlock();
 
-        for (retry = 0; retry < MappedBuffer<MappedBufferT>.NUM_MAX_RETRIEES; ++retry)
+        for (retry = 0; retry < NUM_MAX_RETRIEES; ++retry)
         {
           var bufferSizeBytes = this.BUFFER_SIZE_BYTES;
           // Read current buffer versions.
